[REF] l10n_it_edi: Merge `l10n_it_edi_withholding` module
This commit integrates the functionality from the `l10n_it_edi_withholding` module
into its parent, `l10n_it_edi`.

The separate module was originally created to backport this feature
to a stable version (16.0), making it obsolete now.

Original commit: https://github.com/odoo/odoo/commit/90f9f9f23478bd54625c152c7f5ab2d0c18e3780

task-4988044
