+++ addons/l10n_it_edi/data/template/account.account-it.csv
