+++ addons/l10n_it_edi/models/__init__.py
@@ -1,10 +1,11 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-from . import res_partner
-from . import res_company
-from . import res_config_settings
-from . import account_payment_method_line
+from . import account_chart_template
+from . import account_edi_proxy_user
 from . import account_move
 from . import account_move_send
+from . import account_payment_method_line
+from . import account_tax
 from . import ddt
-from . import account_edi_proxy_user
-from. import l10n_it_document_type
+from . import l10n_it_document_type
+from . import res_company
+from . import res_config_settings
+from . import res_partner
