+++ addons/l10n_it_edi/tests/__init__.py
@@ -3,11 +3,12 @@
 from . import common
 from . import test_account_move_document_type
 from . import test_account_move_payment_method
+from . import test_account_move_send
 from . import test_edi_address
 from . import test_edi_export
 from . import test_edi_import
-from . import test_edi_reverse_charge
 from . import test_edi_pa
-from . import test_account_move_send
-from . import test_res_partner
+from . import test_edi_reverse_charge
 from . import test_it_document_type
+from . import test_res_partner
+from . import test_withholding
