+++ addons/l10n_it_edi/tests/import_xmls/IT00470550013_pfund.xml
