+++ addons/l10n_it_edi/views/l10n_it_view.xml
@@ -86,6 +86,12 @@
                 <field name="l10n_it_edi_attachment_name" optional="hide"/>
                 <field name="l10n_it_edi_state" optional="hide"/>
             </field>
+            <field name="amount_untaxed_in_currency_signed" position="after">
+                <field name="l10n_it_amount_vat_signed" string="VAT" sum="Total" optional="hide"/>
+                <field name="l10n_it_amount_pension_fund_signed" string="Pension Fund" sum="Total" optional="hide"/>
+                <field name="l10n_it_amount_withholding_signed" string="Withholding" sum="Total" optional="hide"/>
+                <field name="l10n_it_amount_before_withholding_signed" string="All Taxes Included" sum="Total" optional="hide"/>
+            </field>
         </field>
     </record>
