--- addons/l10n_it_edi_withholding/data/invoice_it_template.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <template id="account_invoice_it_FatturaPA_export_withholding" inherit_id="l10n_it_edi.account_invoice_it_FatturaPA_export">
-        <xpath expr="//DatiGeneraliDocumento/Numero" position="after">
-            <t t-foreach="withholding_values" t-as="withholding">
-                <DatiRitenuta>
-                    <TipoRitenuta t-esc="format_alphanumeric(withholding['tipo_ritenuta'])"/>
-                    <ImportoRitenuta t-esc="format_float(withholding['importo_ritenuta'], 2)"/>
-                    <AliquotaRitenuta t-esc="format_float(withholding['aliquota_ritenuta'], 2)"/>
-                    <CausalePagamento t-esc="format_alphanumeric(withholding['causale_pagamento'])"/>
-                </DatiRitenuta>
-            </t>
-        </xpath>
-        <xpath expr="//DatiGeneraliDocumento/ImportoTotaleDocumento" position="before">
-            <t t-foreach="pension_fund_values" t-as="pension_fund">
-                <DatiCassaPrevidenziale>
-                    <TipoCassa t-esc="format_alphanumeric(pension_fund['tipo_cassa'])"/>
-                    <AlCassa t-esc="format_float(pension_fund['al_cassa'], 2)"/>
-                    <ImportoContributoCassa t-esc="format_float(pension_fund['importo_contributo_cassa'], 2)"/>
-                    <ImponibileCassa t-esc="format_float(pension_fund['imponibile_cassa'], 2)"/>
-                    <AliquotaIVA t-esc="format_float(pension_fund['aliquota_iva'], 2)"/>
-                    <Ritenuta t-if="pension_fund['ritenuta']"
-                              t-out="pension_fund['ritenuta']"/>
-                    <Natura t-if="pension_fund['natura']"
-                            t-esc="format_alphanumeric(pension_fund['natura'])"/>
-                    <RiferimentoAmministrazione t-if="pension_fund['riferimento_amministrazione']"
-                                                t-esc="format_alphanumeric(pension_fund['riferimento_amministrazione'])"/>
-                </DatiCassaPrevidenziale>
-            </t>
-        </xpath>
-    </template>
-
-</odoo>
