--- addons/l10n_it_edi_withholding/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import account_tax
-from . import account_chart_template
-from . import account_move
